refactor(error): 集約ごとに束ねた thrower を用意し、呼び出し側の対応表引数を無くす
境界のエラー写像で毎回同じ対応表を渡していたのをやめ、集約ごとに対応表を束ねた
throwHttpError を用意して使い回す。

- throw-http-error.ts は対応表を受け取る汎用 throwHttpErrorByTable のみを持つ
- 各集約(auth/article/account-error.ts)が自集約の対応表を束ねた throwHttpError を
  関数宣言で公開する（never 戻り値の関数宣言にすることで呼び出し側の型ナローイングを保つ）
- ハンドラは throwHttpError(error) だけで済む
