+++ application/apps/web/src/server/article/handler/get-articles.ts
@@ -11,8 +11,7 @@ import { offsetPaginationSchema } from '@trend-diary/std/pagination'
 import { z } from 'zod'
 import CONTEXT_KEY from '@/middleware/context'
 import zodValidator, { type ZodValidatedContext } from '@/middleware/zod-validator'
-import { ARTICLE_ERROR_STATUS_TABLE } from '@/server/error/article-error-status'
-import throwHttpError from '@/server/error/throw-http-error'
+import throwHttpError from '@/server/error/article-error'
 import { type ArticleResponse, toArticleResponse } from '../article-response'
 
 const readStatusEnum = z.enum(['0', '1'])
@@ -54,7 +53,7 @@ export default async function getArticles(
     activeUserId,
   )
   if (result.isErr()) {
-    throwHttpError(result.error, ARTICLE_ERROR_STATUS_TABLE)
+    throwHttpError(result.error)
   }
 
   const paginationResult = result.value
