+++ application/apps/web/src/server/article/handler/get-diary.ts
@@ -11,8 +11,7 @@ import { HTTPException } from 'hono/http-exception'
 import { z } from 'zod'
 import CONTEXT_KEY, { mustGet } from '@/middleware/context'
 import zodValidator, { type ZodValidatedContext } from '@/middleware/zod-validator'
-import { ARTICLE_ERROR_STATUS_TABLE } from '@/server/error/article-error-status'
-import throwHttpError from '@/server/error/throw-http-error'
+import throwHttpError from '@/server/error/article-error'
 
 const DATE_STRING_REGEX = /^\d{4}-\d{2}-\d{2}$/
 const DIARY_DATE_MESSAGE = 'date must be a valid JST date'
@@ -95,7 +94,7 @@ export default async function getDiary(c: ZodValidatedContext<[typeof diaryQuery
       DIARY_READ_LIMIT,
     )
     if (detailResult.isErr()) {
-      throwHttpError(detailResult.error, ARTICLE_ERROR_STATUS_TABLE)
+      throwHttpError(detailResult.error)
     }
 
     const response = toDiaryDetailResponse(detailResult.value)
@@ -111,7 +110,7 @@ export default async function getDiary(c: ZodValidatedContext<[typeof diaryQuery
 
   const result = await useCase.getDailyDiaryRange(sessionUser.activeUserId, fromDate, toDate)
   if (result.isErr()) {
-    throwHttpError(result.error, ARTICLE_ERROR_STATUS_TABLE)
+    throwHttpError(result.error)
   }
   const response = toDiaryResponse(result.value)
   logger.info('daily diary range retrieved successfully', {
