+++ application/apps/web/src/server/error/account-error.ts
@@ -1,0 +1,9 @@
+import { ActiveUserNotFoundError } from '@trend-diary/domain/account'
+import throwHttpErrorByTable, { type ErrorStatusTable } from './throw-http-error'
+
+const ERROR_STATUS_TABLE: ErrorStatusTable = [[ActiveUserNotFoundError, 404]]
+
+// アカウント集約のドメインエラーを HTTPException へ写像して送出する。
+export default function throwHttpError(error: Error): never {
+  return throwHttpErrorByTable(error, ERROR_STATUS_TABLE)
+}
