--- application/apps/web/src/server/error/article-error-status.ts
@@ -1,5 +1,0 @@
-import { ArticleNotFoundError } from '@trend-diary/domain/article'
-import type { ErrorStatusTable } from './throw-http-error'
-
-// 記事集約のドメインエラー → HTTP ステータス対応表。
-export const ARTICLE_ERROR_STATUS_TABLE: ErrorStatusTable = [[ArticleNotFoundError, 404]]
