+++ application/apps/web/src/server/oauth/handler/unlink.ts
@@ -1,8 +1,7 @@
 import { authClientConfig, OAuthClient } from '@trend-diary/authentication'
 import { HTTPException } from 'hono/http-exception'
 import type { ZodValidatedContext } from '@/middleware/zod-validator'
-import { AUTH_ERROR_STATUS_TABLE } from '@/server/error/auth-error-status'
-import throwHttpError from '@/server/error/throw-http-error'
+import throwHttpError from '@/server/error/auth-error'
 import type { oauthProviderParamValidator } from '@/server/oauth/schema'
 
 export default async function oauthUnlink(
@@ -12,7 +11,7 @@ export default async function oauthUnlink(
 
   const oauthClient = new OAuthClient(authClientConfig(c))
   const identitiesResult = await oauthClient.listIdentities()
-  if (identitiesResult.isErr()) throwHttpError(identitiesResult.error, AUTH_ERROR_STATUS_TABLE)
+  if (identitiesResult.isErr()) throwHttpError(identitiesResult.error)
 
   const identities = identitiesResult.value
   const target = identities.find((identity) => identity.provider === provider)
@@ -26,7 +25,7 @@ export default async function oauthUnlink(
   }
 
   const unlinkResult = await oauthClient.unlinkIdentity(target)
-  if (unlinkResult.isErr()) throwHttpError(unlinkResult.error, AUTH_ERROR_STATUS_TABLE)
+  if (unlinkResult.isErr()) throwHttpError(unlinkResult.error)
 
   return c.body(null, 204)
 }
