+++ application/apps/web/src/server/oauth/oauth-start.ts
@@ -7,8 +7,7 @@ import {
 import { setCookie } from 'hono/cookie'
 import type { Result } from 'neverthrow'
 import type { ZodValidatedContext } from '@/middleware/zod-validator'
-import { AUTH_ERROR_STATUS_TABLE } from '@/server/error/auth-error-status'
-import throwHttpError from '@/server/error/throw-http-error'
+import throwHttpError from '@/server/error/auth-error'
 import {
   buildOAuthCallbackUrl,
   OAUTH_COOKIE_OPTIONS,
@@ -38,7 +37,7 @@ export function createOAuthStartHandler<
 
     const oauthClient = new OAuthClient(authClientConfig(c))
     const result = await config.start(oauthClient, provider, buildOAuthCallbackUrl(c, provider))
-    if (result.isErr()) throwHttpError(result.error, AUTH_ERROR_STATUS_TABLE)
+    if (result.isErr()) throwHttpError(result.error)
 
     setCookie(c, OAUTH_FLOW_COOKIE, config.flow, OAUTH_COOKIE_OPTIONS)
     config.setRedirectCookie(c)
