+++ application/apps/web/src/server/passkey/handler/authentication-verify.ts
@@ -5,9 +5,8 @@ import { createAccountUseCase } from '@trend-diary/domain/account'
 import { z } from 'zod'
 import CONTEXT_KEY from '@/middleware/context'
 import zodValidator, { type ZodValidatedContext } from '@/middleware/zod-validator'
-import { ACCOUNT_ERROR_STATUS_TABLE } from '@/server/error/account-error-status'
-import { AUTH_ERROR_STATUS_TABLE } from '@/server/error/auth-error-status'
-import throwHttpError from '@/server/error/throw-http-error'
+import throwAccountHttpError from '@/server/error/account-error'
+import throwAuthHttpError from '@/server/error/auth-error'
 
 // 真正性はSupabaseが検証するため中身の妥当性検証はプロバイダに委ね、ここは認証 ceremony 結果を素通しする
 export const passkeyAuthenticationVerifyInputSchema = z.object({
@@ -33,12 +32,12 @@ export default async function passkeyAuthenticationVerify(
     challengeId: valid.challengeId,
     credential: valid.credential,
   })
-  if (userResult.isErr()) throwHttpError(userResult.error, AUTH_ERROR_STATUS_TABLE)
+  if (userResult.isErr()) throwAuthHttpError(userResult.error)
 
   const rdb = getRdbClient(c.env.DB)
   const accountUseCase = createAccountUseCase(rdb)
   const activeUserResult = await accountUseCase.resolveActiveUser(userResult.value.id)
-  if (activeUserResult.isErr()) throwHttpError(activeUserResult.error, ACCOUNT_ERROR_STATUS_TABLE)
+  if (activeUserResult.isErr()) throwAccountHttpError(activeUserResult.error)
 
   logger.info('passkey login success', { activeUserId: activeUserResult.value.activeUserId })
